AU-2354: feature: AU-2354: Moved validateHiddenFields to validation_helpers.
