+++ e2e/utils/validation_helpers.ts
@@ -403,6 +403,29 @@ const navigateAndValidateProfilePage = async (
 }
 
 /**
+ * The validateHiddenFields function.
+ *
+ * This function checks that the passed in items
+ * in itemsToBeHidden are not visible on a given page.
+ * The functionality is used in tests where the value of
+ * field X alters the visibility of field Y.
+ *
+ * @param page
+ *   Page object from Playwright.
+ * @param itemsToBeHidden
+ *   An array of items that should be hidden.
+ * @param formPageKey
+ *   The form page we are on.
+ */
+const validateHiddenFields = async (page: Page, itemsToBeHidden: string[], formPageKey: string) => {
+  for (const hiddenItem of itemsToBeHidden) {
+    const hiddenSelector = `[data-drupal-selector="${hiddenItem}"]`;
+    await expect(page.locator(hiddenSelector), `Field ${hiddenItem} is not hidden on ${formPageKey}.`).not.toBeVisible();
+    logger(`Field ${hiddenItem} is hidden on ${formPageKey}.`)
+  }
+}
+
+/**
  * The MessageType enum.
  *
  * This enum defines the types of messages that
@@ -496,5 +519,6 @@ const logValidationResults = (
 export {
   validateSubmission,
   validateProfileData,
-  validateExistingProfileData
+  validateExistingProfileData,
+  validateHiddenFields,
 }
